requirement.txt
pandas==1.2.4
streamlit==1.4.0
docx2txt==1.0.6
pdfplumber==0.5.25
nltk==3.6.2
wordcloud==1.10.4
matplotlib==3.4.3
plotly==6.7.0